+++ src/cryptography/hazmat/backends/openssl/dsa.py
@@ -162,23 +162,11 @@ class _DSAPrivateKey(object):
         )
 
     def public_key(self):
-        dsa_cdata = self._backend._lib.DSA_new()
+        dsa_cdata = self._backend._lib.DSAparams_dup(self._dsa_cdata)
         self._backend.openssl_assert(dsa_cdata != self._backend._ffi.NULL)
         dsa_cdata = self._backend._ffi.gc(
             dsa_cdata, self._backend._lib.DSA_free
         )
-        p = self._backend._ffi.new("BIGNUM **")
-        q = self._backend._ffi.new("BIGNUM **")
-        g = self._backend._ffi.new("BIGNUM **")
-        self._backend._lib.DSA_get0_pqg(self._dsa_cdata, p, q, g)
-        self._backend.openssl_assert(p[0] != self._backend._ffi.NULL)
-        self._backend.openssl_assert(q[0] != self._backend._ffi.NULL)
-        self._backend.openssl_assert(g[0] != self._backend._ffi.NULL)
-        p_dup = self._backend._lib.BN_dup(p[0])
-        q_dup = self._backend._lib.BN_dup(q[0])
-        g_dup = self._backend._lib.BN_dup(g[0])
-        res = self._backend._lib.DSA_set0_pqg(dsa_cdata, p_dup, q_dup, g_dup)
-        self._backend.openssl_assert(res == 1)
         pub_key = self._backend._ffi.new("BIGNUM **")
         self._backend._lib.DSA_get0_key(
             self._dsa_cdata, pub_key, self._backend._ffi.NULL
@@ -193,23 +181,11 @@ class _DSAPrivateKey(object):
         return _DSAPublicKey(self._backend, dsa_cdata, evp_pkey)
 
     def parameters(self):
-        dsa_cdata = self._backend._lib.DSA_new()
+        dsa_cdata = self._backend._lib.DSAparams_dup(self._dsa_cdata)
         self._backend.openssl_assert(dsa_cdata != self._backend._ffi.NULL)
         dsa_cdata = self._backend._ffi.gc(
             dsa_cdata, self._backend._lib.DSA_free
         )
-        p = self._backend._ffi.new("BIGNUM **")
-        q = self._backend._ffi.new("BIGNUM **")
-        g = self._backend._ffi.new("BIGNUM **")
-        self._backend._lib.DSA_get0_pqg(self._dsa_cdata, p, q, g)
-        self._backend.openssl_assert(p[0] != self._backend._ffi.NULL)
-        self._backend.openssl_assert(q[0] != self._backend._ffi.NULL)
-        self._backend.openssl_assert(g[0] != self._backend._ffi.NULL)
-        p_dup = self._backend._lib.BN_dup(p[0])
-        q_dup = self._backend._lib.BN_dup(q[0])
-        g_dup = self._backend._lib.BN_dup(g[0])
-        res = self._backend._lib.DSA_set0_pqg(dsa_cdata, p_dup, q_dup, g_dup)
-        self._backend.openssl_assert(res == 1)
         return _DSAParameters(self._backend, dsa_cdata)
 
     def private_bytes(self, encoding, format, encryption_algorithm):
@@ -220,6 +196,11 @@ class _DSAPrivateKey(object):
             self._evp_pkey,
             self._dsa_cdata
         )
+
+    def sign(self, data, algorithm):
+        signer = self.signer(algorithm)
+        signer.update(data)
+        return signer.finalize()
 
 
 @utils.register_interface(dsa.DSAPublicKeyWithSerialization)
@@ -268,23 +249,10 @@ class _DSAPublicKey(object):
         )
 
     def parameters(self):
-        dsa_cdata = self._backend._lib.DSA_new()
-        self._backend.openssl_assert(dsa_cdata != self._backend._ffi.NULL)
+        dsa_cdata = self._backend._lib.DSAparams_dup(self._dsa_cdata)
         dsa_cdata = self._backend._ffi.gc(
             dsa_cdata, self._backend._lib.DSA_free
         )
-        p = self._backend._ffi.new("BIGNUM **")
-        q = self._backend._ffi.new("BIGNUM **")
-        g = self._backend._ffi.new("BIGNUM **")
-        self._backend._lib.DSA_get0_pqg(self._dsa_cdata, p, q, g)
-        self._backend.openssl_assert(p[0] != self._backend._ffi.NULL)
-        self._backend.openssl_assert(q[0] != self._backend._ffi.NULL)
-        self._backend.openssl_assert(g[0] != self._backend._ffi.NULL)
-        p_dup = self._backend._lib.BN_dup(p[0])
-        q_dup = self._backend._lib.BN_dup(q[0])
-        g_dup = self._backend._lib.BN_dup(g[0])
-        res = self._backend._lib.DSA_set0_pqg(dsa_cdata, p_dup, q_dup, g_dup)
-        self._backend.openssl_assert(res == 1)
         return _DSAParameters(self._backend, dsa_cdata)
 
     def public_bytes(self, encoding, format):
@@ -300,3 +268,8 @@ class _DSAPublicKey(object):
             self._evp_pkey,
             None
         )
+
+    def verify(self, signature, data, algorithm):
+        verifier = self.verifier(signature, algorithm)
+        verifier.update(data)
+        verifier.verify()